Improve Question generator
